Created new Class Symbol that will replace all Strings.
This is because in Java, all Objects declared with they keyword new are
passed by pointer, so rather than passing a bunch of Strings by value,
we can pass in Objects by Pointer. Doing this allowed the DragonCurve
generation to get from a maximum of 23 to 24 generations before running
out of heap memory.